requirements.txt
openai>=1.40.0
pydantic>=2.8
python-dotenv>=1.0
